requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==40.0.2
Deprecated==1.2.13
frozenlist==1.3.3
gitdb==4.0.10
github==1.2.7
GitPython==3.1.31
idna==3.4
multidict==6.0.4
pycparser==2.21
PyGithub==1.58.1
PyJWT==2.6.0
PyNaCl==1.5.0
requests==2.29.0
smmap==5.0.0
typing_extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
yarl==1.9.2
